requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
alembic==1.6.5
altgraph==0.17
ambition-django-timezone-field==2.0.4
aniso8601==8.1.1
anyio==3.1.0
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
asttokens==2.0.5
async-timeout==3.0.1
atramhasis==0.7.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
Babel==2.9.1
beautifulsoup4==4.9.3
block-disposable-email==2.0.0
boto3==1.17.97
botocore==1.20.97
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==8.0.1
colander==1.8.3
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
daphne==3.0.2
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.4
Django-Abstract-Relations==0.1.4
django-address==0.2.5
django-allauth==0.44.0
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-datetime-widget2==0.9.5
django-disposable-email-checker==2.0.2
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-language-field==0.0.3
django-model-utils==4.1.1
django-multiselectfield==0.1.12
django-pagination==1.0.7
django-post-office==3.5.3
django-redis==5.0.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-safedelete==1.0.0
django-storages==1.11.1
django-timezone-field==4.1.2
django-widget-tweaks==1.4.8
djangochannelsrestframework==0.3.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-queryfields==1.0.0
djangorestframework-simplejwt==4.7.1
dogpile.cache==1.1.3
drf-renderer-xlsx==0.4.0
drf-yasg==1.20.0
Eel==0.14.0
et-xmlfile==1.1.0
executing==0.6.0
factory-boy==3.2.0
Faker==8.8.1
fastapi==0.65.2
fcm-django==0.3.10
filelock==3.0.12
fleming==0.6.0
frozendict==2.0.3
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
graphene==3.0b7
graphene-django==3.0.0b7
graphql-core==3.1.5
graphql-relay==3.1.0
greenlet==1.1.0
gunicorn==20.1.0
h11==0.12.0
h2==4.0.0
hiredis==2.0.0
hpack==4.0.0
html5lib==1.1
httpcore==0.13.6
httpx==0.18.2
hupper==1.10.3
hyperframe==6.0.1
hyperlink==21.0.0
icecream==2.1.1
idna==2.10
importlib-metadata==2.1.1
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
install==1.3.4
iso8601==0.1.14
isodate==0.6.0
isort==5.8.0
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jsonfield==3.1.0
language-tags==1.1.0
lazy-object-proxy==1.4.2
lxml==4.6.3
macholib==1.14
mailslurp-client==12.5.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.2
msgpack-python==0.5.6
multidict==5.1.0
numpy==1.20.3
oauthlib==3.1.1
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
PasteDeploy==2.1.1
path==16.0.0
pathlib==1.0.1
pbr==5.6.0
pep8==1.7.1
Pillow==8.2.0
pipenv==2021.5.29
plaster==1.0
plaster-pastedeploy==0.7
pluggy==0.13.1
promise==2.3
psycopg2==2.9.1
psycopg2-binary==2.9.1
psycopg2-connection==0.1.1.1
pure-eval==0.2.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyfcm==1.4.9
Pygments==2.9.0
pyinstaller-hooks-contrib==2021.1
PyJWT==2.1.0
PyLD==2.0.3
pylint==2.8.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyramid==2.0
pyramid-jinja2==2.8
pyramid-rewrite==0.2
pyramid-skosprovider==0.9.2
pyramid-tm==2.4
pytest==6.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
rdflib==5.0.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
rfc3987==1.3.8
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
Rx==1.6.1
s3transfer==0.4.2
selenium==3.141.0
service-identity==21.1.0
simplejson==3.17.2
singledispatch==3.6.2
six==1.16.0
skosprovider==0.7.1
skosprovider-getty==0.5.1
skosprovider-rdf==0.8.1
skosprovider-sqlalchemy==0.6.0
SMTPEmail==0.4.2
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.4.19
sqlparse==0.4.1
starlette==0.14.2
stevedore==3.3.0
temp-mail==0.2
tempMail2==1.0.2
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
transaction==3.0.1
translationstring==1.4
Twisted==21.2.0
txaio==21.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
varname==0.6.5
venusian==3.0.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
waitress==2.0.0
webencodings==0.5.1
WebOb==1.8.7
websockets==9.1
Werkzeug==2.0.1
whichcraft==0.6.1
whitenoise==5.2.0
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.interface==5.4.0
zope.sqlalchemy==1.4

celery>=5.1.2